Pull-up: Helper to BaseHelper
