+++ test/Peach/Markup/ContainerTestImpl.php
@@ -60,7 +60,7 @@ class ContainerTestImpl
     {
         $test   = $this->test;
         $obj    = $this->object;
-        $helper = new Helper(new DefaultBuilder());
+        $helper = new BaseHelper(new DefaultBuilder());
         $node1  = $helper->createObject("div")->append("Test 1")->append("Test 2")->append("Test 3");
         $obj->appendNode($node1);
